requirements.txt
Django==1.11
django-filter==0.13.0
django-js-reverse==0.7.0
django-webpack-loader==0.3.0
djangorestframework==3.3.3
Markdown==2.6.6

psycopg2==2.7.1
djangorestframework==3.6.2
